requirements.txt
asgiref==3.5.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.13
django-import-export==2.8.0
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
pkg-resources==0.0.0
pytz==2022.1
PyYAML==6.0
sqlparse==0.4.2
tablib==3.2.1
typing-extensions==4.2.0
xlrd==2.0.1
xlwt==1.3.0
